+++ app/client/packages/design-system/ads/src/DismissibleTab/DismissibleTab.types.ts
@@ -1,0 +1,10 @@
+import type React from "react";
+
+export interface DismissibleTabProps {
+  children: React.ReactNode;
+  dataTestId?: string;
+  isActive: boolean;
+  onClick: () => void;
+  onClose: (e: React.MouseEvent) => void;
+  onDoubleClick?: () => void;
+}
